+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.3.10 (Mar 24, 2025)
+
+- Added route to `SectionGridView::entriesCountColumn()`
+- Enhanced `SectionEntryController::actionIndex()` to redirect to the default type if no type is set
+- Enhanced `SectionLinkedEntryGridView` to link to the correct entry type if set
+- Enhanced `SectionEntryGridView` to filter out unsupported entry types
+
 ## 2.3.9 (Mar 20, 2025)
 
 - Fixed an issue where the entry slug database index was not created during migration
@@ -308,7 +315,7 @@
 
 ## v2.0.6 (Nov 7, 2023)
 
-- Added default view path to `Sections` widget
+- Added the default view path to `Sections` widget
 - Enhanced `Entry::getRoute()`, it now also returns the route when it has descendants
 - Enhanced `EntryParentIdFieldTrait` to truncate long parent slugs